scraper/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
requests==2.31.0
beautifulsoup4==4.12.3
firebase-admin==6.4.0
python-dotenv==1.0.0
lxml==5.2.1
groq==0.11.0
